moved Trainer to its separate file
